Iteration 15: Diffusion
Diffusion algorithm has reached critical mass.  There is one thing left to try, but it can wait till later.  I will need to remember to look at the diffusion bias again.  There was a lot of interesting statistical properties which can be leveraged to control the behavior of the simulation.